Move test to test_mixins.py
